+++ src/hooks/useApi.ts
@@ -3,6 +3,7 @@ import { APIContext } from "../contexts/DataContext";
 import { Product } from "../types/productType";
 import postCartItems from "../api/postCartItems";
 import patchCartItemQuantity from "../api/patchCartItemQuantity";
+import deleteCartItems from "../api/deleteCartItems";
 
 const CART_MAX_COUNT = 50;
 
@@ -77,10 +78,22 @@ export function useAPI<T>({
     setErrorMessage(error.message);
   };
 
+  const removeFromCart = async (productId: number) => {
+    const { error } = await deleteCartItems(productId);
+
+    if (!error?.message) {
+      setErrorMessage("");
+      return request();
+    }
+
+    setErrorMessage(error.message);
+  };
+
   return {
     data: data[name] as T | undefined,
     refetch: request,
     addToCart,
     patchQuantity,
+    removeFromCart,
   };
 }
